refactor(suite-native): confirmAddressOnDevice
- device address confirmation logic moved from `packages/suite` to `suite-common/wallet-utils`
